AIOutputValidation/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
gitpython==3.1.43
playwright==1.44.0
httpx==0.27.0
ollama==0.2.1
